Moved the AlarmReminderItem to the AlarmReminder
